cloudlet-gateway/requirements/dev.txt
alembic==0.9.1
aniso8601==1.2.0
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.3
beautifulsoup4==4.5.3
blinker==1.4
cached-property==1.3.0
cffi==1.10.0
click==6.7
cliff==2.5.0
cliff-tablib==1.1
cmd2==0.7.0
colorama==0.3.7
configparser==3.5.0
cryptography==1.8.1
cssmin==0.2.0
debtcollector==1.13.0
docker==2.2.1
docker-compose==1.12.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
factory-boy==2.8.1
Faker==0.7.11
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
flake8-docstrings==1.0.3
flake8-isort==2.2
flake8-polyfill==1.0.1
flake8-quotes==0.9.0
Flask==0.12
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Caching==1.2.0
Flask-DebugToolbar==0.10.1
Flask-JWT-Extended==1.4.0
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.1.1
gunicorn==19.7.1
idna==2.5
ipaddress==1.0.18
iso8601==0.1.11
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
jsmin==2.2.1
jsonpatch==1.15
jsonpointer==1.10
jsonschema==2.6.0
keystoneauth1==2.19.0
Mako==1.0.6
MarkupSafe==1.0
mccabe==0.6.1
monotonic==1.3
msgpack-python==0.4.8
netaddr==0.7.19
netifaces==0.10.5
os-client-config==1.26.0
osc-lib==1.3.0
oslo.config==1.9.3
oslo.i18n==3.15.0
oslo.serialization==2.18.0
oslo.utils==3.25.0
packaging==16.8
pbr==2.1.0
pep8-naming==0.4.1
positional==1.1.1
prettytable==0.7.2
psycopg2==2.7.1
py==1.4.33
pycodestyle==2.3.1
pycparser==2.17
pydocstyle==1.1.1
pyflakes==1.5.0
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyparsing==2.2.0
pytest==3.0.7
python-cinderclient==1.1.3
python-dateutil==2.6.0
python-editor==1.0.3
python-glanceclient==0.17.3
python-heatclient==1.9.0
python-keystoneclient==1.3.4
python-neutronclient==2.4.0
python-novaclient==7.0.0
python-swiftclient==3.3.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.12.5
requestsexceptions==1.2.0
rfc3986==0.4.1
ruamel.ordereddict==0.4.9
ruamel.yaml==0.14.6
simplejson==3.10.0
simplekv==0.10.0
six==1.10.0
SQLAlchemy==1.1.5
stevedore==1.21.0
tablib==0.11.4
testfixtures==4.13.5
texttable==0.8.8
typing==3.6.1
unicodecsv==0.14.1
waitress==1.0.2
warlock==1.3.0
webassets==0.12.1
WebOb==1.7.2
websocket-client==0.40.0
WebTest==2.0.27
Werkzeug==0.12.1
wrapt==1.10.10
WTForms==2.1
celery==4.0.2
